+++ src/ui/buttons/buttonDesigns/switchLogoAndShowLabelText.jsx
